pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-ontology-agent"
version = "0.1.0"
description = "基于本地 LLM (Ollama) 的多智能体领域 Ontology 抽取与推理系统"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Jiangsheng Yu"},
]
maintainers = [
    {name = "Jiangsheng Yu"},
]
keywords = [
    "LLM", "ontology", "OWL", "SHACL", "SPARQL",
    "knowledge-graph", "multi-agent", "reasoning", "Ollama",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "rdflib>=7.0",
    "owlrl>=6.0",
    "pyshacl>=0.25",
    "python-docx>=1.0",
    "flask>=3.0",
]

[project.optional-dependencies]
pdf = ["PyMuPDF>=1.23"]
dev = ["pytest>=7.0", "ruff>=0.4"]

[project.scripts]
ontology-agent = "main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
